* Fix: Table Of Contents - Removed the dependancy on data attribues for header.
