+++ src/blocks/table-of-contents/save.js
@@ -3,9 +3,8 @@
  */
 
 import classnames from "classnames"
-import UAGB_Block_Icons from "../../../dist/blocks/uagb-controls/block-icons"
 import renderSVG from "../../../dist/blocks/uagb-controls/renderIcon"
-import TOC from './table-of-contents';
+import TableOfContents from './toc';
 
 const { __ } = wp.i18n;
 
@@ -24,7 +23,6 @@ export default function save( props ) {
 		heading,
 		headerLinks,
 		mappingHeaders,
-		scrollToTop,
 		smoothScroll,
 		makeCollapsible,
 		icon,
@@ -44,15 +42,15 @@ export default function save( props ) {
 	return (
 
 		<div className={ classnames(
-			className,
-			`uagb-toc__align-${align}`,
-			`uagb-toc__columns-${tColumns}`,
-			( initialCollapse ) ? `uagb-toc__collapse` : '',
-			`uagb-block-${ block_id }`
-		) }
-		data-scroll={smoothScroll}
-		data-offset={smoothScrollOffset}
-		data-delay={smoothScrollDelay}
+				className,
+				`uagb-toc__align-${align}`,
+				`uagb-toc__columns-${tColumns}`,
+				( initialCollapse ) ? `uagb-toc__collapse` : '',
+				`uagb-block-${ block_id }`
+			) }
+			data-scroll={smoothScroll}
+			data-offset={smoothScrollOffset}
+			data-delay={smoothScrollDelay}
 		>
 			<div className="uagb-toc__wrap">
 				<div className="uagb-toc__title-wrap">
@@ -63,7 +61,7 @@ export default function save( props ) {
 					/>
 					{icon_html}
 				</div>
-				<TOC
+				<TableOfContents
 					mappingHeaders={mappingHeaders}
 					headers={headerLinks && JSON.parse(headerLinks)}
 				/>
